add signup endpoint
