+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "python.pythonPath": "C:\\Users\\WOHU-DENNIS 001\\.virtualenvs\\practice2-PzG_ZHIp\\Scripts\\python.exe"
+    "python.pythonPath": "C:\\Users\\WOHU-DENNIS 001\\.virtualenvs\\My_Diary-03xE5-M5\\Scripts\\python.exe"
 }
