--- tests/test_user.py
@@ -1,6 +1,0 @@
-from api.v1.user import User
-
-
-def test_add_user():
-    person = User('soko', "00", "00")
-    assert person.get_user().name == 'soko'
